rename avl_tree to avl_map
