+++ include/mctp/util/avl_map/decl.h
@@ -1,5 +1,5 @@
-#ifndef _MCTP_UTIL_AVL_TREE_DECL_H_
-#define _MCTP_UTIL_AVL_TREE_DECL_H_
+#ifndef _MCTP_UTIL_AVL_MAP_DECL_H_
+#define _MCTP_UTIL_AVL_MAP_DECL_H_
 
 #include <mctp/util/typename.h>
 #include <mctp/util/extern_c.h>
@@ -7,7 +7,7 @@
 
 typedef int8_t avl_height_t;
 
-#define _x_avl_tree_type_decl_private(  \
+#define _x_avl_map_type_decl_private(   \
     typename,                           \
     avl_type_t,                         \
     avl_node_t,                         \
@@ -57,12 +57,12 @@ _x_method(typename, get)(               \
 EXTERN_C_END
 
 
-#define _x_avl_tree_type_decl(  \
+#define _x_avl_map_type_decl(   \
     typename,                   \
     _x_value_t,                 \
     _x_key_t                    \
 )                               \
-_x_avl_tree_type_decl_private(  \
+_x_avl_map_type_decl_private(   \
     typename,                   \
     _x_type_t(typename),        \
     _x_node_t(typename),        \
@@ -70,4 +70,4 @@ _x_avl_tree_type_decl_private(  \
     _x_key_t                    \
 )
 
-#endif // _MCTP_UTIL_AVL_TREE_DECL_H_
+#endif // _MCTP_UTIL_AVL_MAP_DECL_H_
