+++ src/packet_queue_map.c
@@ -1,5 +1,5 @@
 #include <mctp/core/packet_queue_map.h>
-#include <mctp/util/avl_tree/impl.h>
+#include <mctp/util/avl_map/impl.h>
 
 typedef mctp_pktq_t avl_value_t;
 typedef mctp_msg_ctx_t avl_key_t;
@@ -32,7 +32,7 @@ static bool avl_value_eq(
 }
 
 
-_x_avl_tree_value_iface(
+_x_avl_map_value_iface(
     mctp_pktq_map,
     avl_value_t,
     avl_key_t
@@ -43,7 +43,7 @@ _x_avl_tree_value_iface(
     avl_value_eq
 )
 
-_x_avl_tree_type_impl(
+_x_avl_map_type_impl(
     mctp_pktq_map,
     avl_value_t,
     avl_key_t
